Fix warnings in AVI
git-svn-id: https://svn.ices.utexas.edu/repos/iss/projects/GaloisCpp/trunk@11154 522c16a8-e32f-0410-b4cc-8071c0b41197

+++ apps/avi/main/AVIodgAuto.h
@@ -84,7 +84,7 @@ protected:
     vertices.clear ();
     vtxPtrs.clear ();
 
-    for (size_t i = 0; i < meshInit.getNumNodes (); ++i) {
+    for (int i = 0; i < meshInit.getNumNodes (); ++i) {
       vertices.push_back (NItem (i));
     }
